app/requirements.txt
appdirs==1.4.3
click==6.7
DAWG-Python==0.7.2
Django==1.11
django-rq==0.9.5
djangorestframework==3.6.2
docopt==0.6.2
packaging==16.8
psycopg2==2.7.1
pymorphy2==0.8
pymorphy2-dicts==2.4.393442.3710985
pyparsing==2.2.0
pytz==2017.2
redis==2.10.5
rq==0.7.1
six==1.10.0
Pillow==5.4.1
